+++ smx/sourcemod/natives/string.py
@@ -1,6 +1,15 @@
 from __future__ import annotations
 
+import re
+
 from smx.sourcemod.natives.base import native, SourceModNativesMixin, WritableString
+
+
+RGX_BREAK_STRING = re.compile(
+    r'^\s*(?:"(?P<quoted>[^"]*)(?:"|$)|(?P<unquoted>\S+))(?P<end_ws>\s*)',
+    # ASCII mode ensures \s only matches "\n\v\r\t\f ", like SM does
+    flags=re.ASCII,
+)
 
 
 class StringNatives(SourceModNativesMixin):
@@ -45,3 +54,21 @@ class StringNatives(SourceModNativesMixin):
 
         buf.max_length = len(string)
         return buf.write(string.strip(), null_terminate=True)
+
+    @native('string', 'writable_string')
+    def BreakString(self, source: str, arg: WritableString) -> int:
+        if not source:
+            arg.write(b'', null_terminate=True)
+            return -1
+
+        match = RGX_BREAK_STRING.match(source)
+        if match is None:
+            arg.write(b'', null_terminate=True)
+            return -1
+
+        arg.write(match.group('quoted') or match.group('unquoted'), null_terminate=True)
+
+        if not match.group('end_ws'):
+            return -1
+
+        return match.endpos